[01130] Extract shared MockState<T> test helper to eliminate duplication
