+++ src/Ivy.Test/DictationTests.cs
@@ -1,6 +1,4 @@
 using System.Net;
-using System.Reactive.Subjects;
-using System.Runtime.CompilerServices;
 
 namespace Ivy.Test;
 
@@ -83,30 +81,6 @@ public class DictationTests
     public void AzureSpeechTranscriptionService_Constructor_ThrowsOnEmptyKey()
     {
         Assert.Throws<ArgumentException>(() => new AzureSpeechTranscriptionService("region", ""));
-    }
-
-    private class MockState<T>(T value) : IState<T>
-    {
-        private readonly Subject<T> _subject = new();
-        public T Value { get; set; } = value;
-
-        [OverloadResolutionPriority(1)]
-        public T Set(T value) { Value = value; return Value; }
-        public T Set(Func<T, T> setter) { Value = setter(Value); return Value; }
-        public T Reset() => Set(default(T)!);
-        public Type GetStateType() => typeof(T);
-
-        public IDisposable Subscribe(IObserver<T> observer)
-        {
-            observer.OnNext(Value);
-            return _subject.Subscribe(observer);
-        }
-
-        public void Dispose() => _subject.Dispose();
-        public IDisposable SubscribeAny(Action action) => _subject.Subscribe(_ => action());
-        public IDisposable SubscribeAny(Action<object?> action) => _subject.Subscribe(x => action(x));
-        public IEffectTrigger ToTrigger() => EffectTrigger.OnStateChange(this);
-        public object? GetValueAsObject() => Value;
     }
 
     private class MockHttpHandler(string responseContent, HttpStatusCode statusCode) : HttpMessageHandler
