+++ src/Ivy.Test/TextInputTests.cs
@@ -1,6 +1,3 @@
-using System.Reactive.Subjects;
-using System.Runtime.CompilerServices;
-
 namespace Ivy.Test;
 
 //Todo: I don't really think we need these tests, as if TextInput<string> works, then TextInput should work too.
@@ -228,58 +225,4 @@ public class TextInputTests
         Assert.Null(nullableStringInput.Value);
     }
 
-    private class MockState<T>(T value) : IState<T>
-    {
-        private readonly Subject<T> _subject = new();
-
-        public T Value { get; set; } = value;
-
-        [OverloadResolutionPriority(1)]
-        public T Set(T value)
-        {
-            Value = value;
-            return Value;
-        }
-
-        public T Set(Func<T, T> setter)
-        {
-            Value = setter(Value);
-            return Value;
-        }
-
-        public T Reset()
-        {
-            return Set(default(T)!);
-        }
-
-        public Type GetStateType() => typeof(T);
-
-        public IDisposable Subscribe(IObserver<T> observer)
-        {
-            observer.OnNext(Value);
-            return _subject.Subscribe(observer);
-        }
-
-        public void Dispose()
-        {
-            _subject.Dispose();
-        }
-
-        public IDisposable SubscribeAny(Action action)
-        {
-            return _subject.Subscribe(_ => action());
-        }
-
-        public IDisposable SubscribeAny(Action<object?> action)
-        {
-            return _subject.Subscribe(x => action(x));
-        }
-
-        public IEffectTrigger ToTrigger()
-        {
-            return EffectTrigger.OnStateChange(this);
-        }
-
-        public object? GetValueAsObject() => Value;
-    }
 }
